[스프링 - 협업 미션] 바다(조아라) 미션 제출합니다. (#28)
* build: mysql 의존성 추가

* build: 설정파일 분리

* build: yml 설정 추가

* build: yml 설정 변경

* feat: 전체 노선 목록 상세조회를 위한 dto 추가

* feat: 전체 노선 목록 상세조회 API 매핑

* feat: 노선의 총 길이 구하는 기능 구현 및 테스트코드 작성

* refactor: 자동 포매팅

* feat(DataLoader): db에 데이터가 없을 때만 초기데이터를 추가하도록 수정

* build: javax 의존성 추가

* docs(README): 3,4단계 기능구현 목록 작성

* feat: 요금 계산 기능 구현 및 테스트코드 작성

* feat: 경로조회 결과에 요금 정보 포함 및 인수테스트 수정

* refactor(LineTest): 테스트에서 Fixture를 사용하도록 수정

* refactor: 삭제응답 200 -> 204

* refactor: 서블릿 인터셉터 설정 및 token 검증

* feat: GlobalException 설정

* feat: station 예외 로직 구현

* feat: line 예외 로직 구현

* feat: Station 예외 처리 로직

* feat: Line 예외 처리 로직

* feat: Path 예외 처리 로직

* refactor: Name Validation 수정

* refactor: Auth 예외 처리 로직

* feat: 노선에 추가요금 추가, LINE 테이블에 추가요금 컬럼 추가

* feat: 중복 이메일 검증 로직 추가

* feat: json 형태로 들어오지 않았을 때 415예외 던지도록 핸들러 추가

* feat: CORS 설정 추가

* feat(SubwayPath): 환승시 추가요금 계산로직 구현

* refactor(StationRequest): 역이름 길이제한 3 -> 2로 수정

* refactor(StationDao): 역 조회 시 id 순으로 조회되도록 수정

* refactor(schema.sql): color 필드 길이제한 수정

* refactor(LineRequest): 노선이름 길이제한 3 -> 2로 수정

* refactor(schema.sql): 노선 삭제 시 포함된 구간 같이 삭제되도록 sql 옵션 추가

* feat: 연령별 할인 정책 추가

* refactor: 구간 생성 시 201 반환하도록 수정

* refactor: 모든 컨트롤러에 API response타입 명시

* build: 빌드 설정 중 local 프로파일 수정

* refactor: 불필요한 주석 삭제

* build: local 프로파일 설정 변경

* refactor: 거리에 따른 요금 계산과 나이에 따른 요금 계산 로직 수정

* refactor: 전체적으로 코드 정리
- final 접근제한자 추가
- 람다 사용하는 부분 메서드 참조로 수정
- 가독성을 위해 개행 추가
- raw 타입 제네릭 사용
- 생성자 체이닝

* refactor: exception과 config 패키지 정리

* refactor: RequestParam을 ModelAttribute 이용해 수정

* refactor: 서비스에서 검증 메서드 분리

* refactor: 테스트 코드 DisplayName 좀 더 명확하게 수정

* refactor: 테스트코드 리팩터링

* refactor(Sections): 중복 검증 메서드 분리

* refactor: 불필요한 주석 제거

* refactor: Dao에서 불필요한 코드 정리

* refactor: Service단 트랜잭션 처리

* refactor(Fare): 클래스 변수와 인스턴스 변수 사이 개행 추가

* refactor(Station): 중복코드 삭제

* refactor(schema.sql): color 컬럼에 unique 설정 추가

* refactor(DistanceStrategy): 정수 나누기 연산에 소수점 사라지지 않도록 수정

* refactor(MemberDao): findByEmail이 Optional을 반환하도록 수정

* refactor(GlobalExceptionHandler): 커스텀익셉션의 로그레벨 Error -> Warn으로 수정

* refactor: 거리 요금 정책에 책임 연쇄 패턴 적용

* refactor(SubwayException): 상속으로만 사용하는 SubwayException에 abstract 추가

* refactor(SubwayException): SubwayException에 status 추가

* refactor(StationService): 역 삭제 예외처리 수정

* refactor(AuthorizationException): SubwayException 상속하도록 수정

* refactor: DefaultChain Bean등록 후 Config에서 체이닝작업 하도록 수정

* refactor: Transactional 필요 없는 부분 삭제

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,6 +1,8 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.exception.DuplicateException;
+import wooteco.subway.exception.InvalidInputException;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
@@ -11,19 +13,28 @@ import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+
+    private final StationDao stationDao;
 
     public StationService(StationDao stationDao) {
         this.stationDao = stationDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
+        validateDuplicateName(stationRequest.getName());
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
+    private void validateDuplicateName(String name) {
+        if (stationDao.existsByName(name)) {
+            throw new DuplicateException("이미 존재하는 역 이름 입니다. (입력된 이름 값 : " + name + ")");
+        }
+    }
+
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id)
+                .orElseThrow(() -> new InvalidInputException("해당하는 id의 역이 없습니다."));
     }
 
     public List<StationResponse> findAllStationResponses() {
@@ -35,6 +46,14 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
+        findStationById(id);
+        validateDeletable(id);
         stationDao.deleteById(id);
+    }
+
+    private void validateDeletable(Long id) {
+        if (stationDao.countRegisteredById(id) != 0) {
+            throw new InvalidInputException("노선에 등록된 역은 삭제할 수 없습니다.");
+        }
     }
 }
